After TomlParser::parse() is done, before returning the doc object, close the file.
This is mostly required on Windows due to
the fact that when a file is opened it is locked from suppression/renaming
by other processes. If a TomlParser object is not freed then the toml file
becomes actually locked.

To prevent this, explicitely close the toml file after beeing read/parsed.